AcpiHelp: Fix a printf formatting warning under Linux.

+++ source/tools/acpihelp/ahmain.c
@@ -452,7 +452,7 @@ AhDisplayOperator (
             Position = TokenLength;
         }
 
-        printf ("%.*s ", TokenLength, This);
+        printf ("%.*s ", (int) TokenLength, This);
         This = Next + 1;
     }
